+++ packages/frontend/app/(dashboard)/course/[cid]/(settings)/settings/lms_integrations/components/UpsertIntegrationModal.tsx
@@ -5,8 +5,9 @@ import { useEffect, useMemo, useState } from 'react'
 import {
   LMSApiResponseStatus,
   LMSCourseIntegrationPartial,
-  LMSIntegration,
+  LMSIntegrationPlatform,
   LMSOrganizationIntegrationPartial,
+  UpsertLMSCourseParams,
 } from '@koh/common'
 import { API } from '@/app/api'
 import { BaseOptionType } from 'antd/es/select'
@@ -26,7 +27,7 @@ type CreateIntegrationModalProps = {
   testLMSConnection: (
     key: string,
     course: string,
-    platform: LMSIntegration,
+    platform: LMSIntegrationPlatform,
   ) => Promise<LMSApiResponseStatus>
   onCreate: () => void
 }
@@ -47,7 +48,7 @@ const UpsertIntegrationModal: React.FC<CreateIntegrationModalProps> = ({
   const [apiKeyExpiry, setApiKeyExpiry] = useState<Date | undefined>(undefined)
   const [apiCourseId, setApiCourseId] = useState<string>('')
   const [selectedPlatform, setSelectedPlatform] = useState<
-    LMSIntegration | undefined
+    LMSIntegrationPlatform | undefined
   >(baseIntegration?.apiPlatform)
 
   const usePlatform = useMemo(() => {
@@ -55,13 +56,14 @@ const UpsertIntegrationModal: React.FC<CreateIntegrationModalProps> = ({
       return selectedPlatform ?? baseIntegration.apiPlatform
     else if (selectedIntegration != undefined)
       return selectedIntegration.apiPlatform
-    else return 'None' as LMSIntegration
+    else return 'None' as LMSIntegrationPlatform
   }, [baseIntegration, selectedIntegration, selectedPlatform])
 
   const mappedLMS = useMemo(() => {
     const pairs: { [key: string]: string } = {}
-    Object.keys(LMSIntegration).map((integration: string) => {
-      pairs[integration] = LMSIntegration[integration as LMSIntegration]
+    Object.keys(LMSIntegrationPlatform).map((integration: string) => {
+      pairs[integration] =
+        LMSIntegrationPlatform[integration as LMSIntegrationPlatform]
     })
     return pairs
   }, [])
@@ -88,18 +90,18 @@ const UpsertIntegrationModal: React.FC<CreateIntegrationModalProps> = ({
 
     testLMSConnection(apiKey, apiCourseId, usePlatform).then((result) => {
       if (result == LMSApiResponseStatus.Success) {
-        const body: { [key: string]: any } = {
+        const body: UpsertLMSCourseParams = {
           apiPlatform: usePlatform,
           apiKey,
           apiKeyExpiry,
           apiCourseId,
         }
         if (baseIntegration != undefined) {
-          body['apiKeyExpiryDeleted'] = apiKeyExpiry == undefined
+          body.apiKeyExpiryDeleted = apiKeyExpiry == undefined
         }
 
-        API.course
-          .upsertIntegration(courseId, body as any)
+        API.lmsIntegration
+          .upsertCourseIntegration(courseId, body)
           .then((result) => {
             if (!result) {
               message.error(