+++ packages/server/ormconfig.ts
@@ -35,6 +35,8 @@ import { QueueInviteModel } from './src/queue/queue-invite.entity';
 import { LMSOrganizationIntegrationModel } from './src/lmsIntegration/lmsOrgIntegration.entity';
 import { LMSCourseIntegrationModel } from './src/lmsIntegration/lmsCourseIntegration.entity';
 import { LMSAssignmentModel } from './src/lmsIntegration/lmsAssignment.entity';
+import { LMSAnnouncementModel } from './src/lmsIntegration/lmsAnnouncement.entity';
+import { UnreadAsyncQuestionModel } from './src/asyncQuestion/unread-async-question.entity';
 import { AsyncQuestionCommentModel } from './src/asyncQuestion/asyncQuestionComment.entity';
 
 config();
@@ -91,6 +93,8 @@ const typeorm = {
     LMSOrganizationIntegrationModel,
     LMSCourseIntegrationModel,
     LMSAssignmentModel,
+    UnreadAsyncQuestionModel,
+    LMSAnnouncementModel,
   ],
   keepConnectionAlive: true,
   logging: