+++ packages/server/src/course/course.controller.ts
@@ -1,5 +1,4 @@
 import {
-  CoursePartial,
   CourseSettingsRequestBody,
   CourseSettingsResponse,
   EditCourseInfoParams,
@@ -7,7 +6,6 @@ import {
   GetCourseResponse,
   GetCourseUserInfoResponse,
   GetLimitedCourseResponse,
-  LMSCourseIntegrationPartial,
   QuestionStatusKeys,
   QueueConfig,
   QueueInvite,
@@ -40,6 +38,7 @@ import {
   UseGuards,
   UseInterceptors,
   ParseIntPipe,
+  ForbiddenException,
 } from '@nestjs/common';
 import async from 'async';
 import { Response, Request } from 'express';
@@ -63,9 +62,8 @@ import { ConfigService } from '@nestjs/config';
 import { ApplicationConfigService } from '../config/application_config.service';
 import { getManager } from 'typeorm';
 import { QuestionTypeModel } from 'questionType/question-type.entity';
-import { LMSOrganizationIntegrationModel } from '../lmsIntegration/lmsOrgIntegration.entity';
-import { LMSCourseIntegrationModel } from '../lmsIntegration/lmsCourseIntegration.entity';
 import { QueueCleanService } from 'queue/queue-clean/queue-clean.service';
+import { UnreadAsyncQuestionModel } from 'asyncQuestion/unread-async-question.entity';
 
 @Controller('courses')
 @UseInterceptors(ClassSerializerInterceptor)
@@ -1023,131 +1021,35 @@ export class CourseController {
     return;
   }
 
+  // Moved from userInfo context endpoint as this updates too frequently to make sense caching it with userInfo data
+  @Get(':id/unread_async_count')
+  @UseGuards(JwtAuthGuard)
+  async getUnreadAsyncCount(
+    @Param('id', ParseIntPipe) courseId: number,
+    @UserId() userId: number,
+  ): Promise<number> {
+    const count = await UnreadAsyncQuestionModel.count({
+      where: {
+        userId,
+        courseId,
+        readLatest: false,
+      },
+    });
+
+    return count;
+  }
+
   @Patch(':id/unread_async_count')
   @UseGuards(JwtAuthGuard)
   async updateUnreadAsyncCount(
     @Param('id', ParseIntPipe) courseId: number,
-    @User() user: UserModel,
+    @UserId() userId: number,
   ): Promise<void> {
-    const userCourse = await UserCourseModel.findOne({
-      where: {
-        user,
-        courseId,
-      },
-    });
-
-    if (!userCourse) {
-      throw new NotFoundException('UserCourse not found');
-    }
-
-    userCourse.unreadAsyncQuestions = 0;
-    await userCourse.save();
+    await UnreadAsyncQuestionModel.update(
+      { userId, courseId },
+      { readLatest: true },
+    );
 
     return;
-  }
-
-  @Get(':id/lms_integration')
-  @UseGuards(JwtAuthGuard, CourseRolesGuard)
-  @Roles(Role.PROFESSOR)
-  async getLmsIntegration(
-    @Param('id', ParseIntPipe) courseId: number,
-  ): Promise<LMSCourseIntegrationPartial | undefined> {
-    const lmsIntegration = await LMSCourseIntegrationModel.findOne({
-      where: { courseId: courseId },
-      relations: ['orgIntegration', 'course'],
-    });
-    if (lmsIntegration == undefined) return undefined;
-
-    return {
-      apiPlatform: lmsIntegration.orgIntegration.apiPlatform,
-      courseId: lmsIntegration.courseId,
-      course: {
-        id: lmsIntegration.courseId,
-        name: lmsIntegration.course.name,
-      } satisfies CoursePartial,
-      apiCourseId: lmsIntegration.apiCourseId,
-      apiKeyExpiry: lmsIntegration.apiKeyExpiry,
-    } satisfies LMSCourseIntegrationPartial;
-  }
-
-  @Post(':id/lms_integration/upsert')
-  @UseGuards(JwtAuthGuard, CourseRolesGuard)
-  @Roles(Role.PROFESSOR)
-  async upsertLMSIntegration(
-    @Param('id', ParseIntPipe) courseId: number,
-    @Body() props: any,
-  ): Promise<any> {
-    const orgCourse = await OrganizationCourseModel.findOne({
-      courseId: courseId,
-    });
-    if (!orgCourse) {
-      return ERROR_MESSAGES.courseController.organizationNotFound;
-    }
-
-    const orgIntegration = await LMSOrganizationIntegrationModel.findOne({
-      where: {
-        organizationId: orgCourse.organizationId,
-        apiPlatform: props.apiPlatform,
-      },
-    });
-    if (!orgIntegration) {
-      return ERROR_MESSAGES.courseController.orgIntegrationNotFound;
-    }
-
-    const courseIntegration = await LMSCourseIntegrationModel.findOne({
-      where: { courseId: courseId },
-    });
-
-    if (courseIntegration != undefined) {
-      return await this.courseService.updateLMSIntegration(
-        courseIntegration,
-        orgIntegration,
-        props.apiKeyExpiryDeleted,
-        props.apiCourseId,
-        props.apiKey,
-        props.apiKeyExpiry,
-      );
-    } else {
-      return await this.courseService.createLMSIntegration(
-        orgIntegration,
-        courseId,
-        props.apiCourseId,
-        props.apiKey,
-        props.apiKeyExpiry,
-      );
-    }
-  }
-
-  @Delete(':id/lms_integration/remove')
-  @UseGuards(JwtAuthGuard, CourseRolesGuard)
-  @Roles(Role.PROFESSOR)
-  async removeLMSIntegration(
-    @Param('id', ParseIntPipe) courseId: number,
-    @Body() props: any,
-  ): Promise<any> {
-    const orgCourse = await OrganizationCourseModel.findOne({
-      courseId: courseId,
-    });
-    if (!orgCourse) {
-      return ERROR_MESSAGES.courseController.organizationNotFound;
-    }
-
-    const orgIntegration = await LMSOrganizationIntegrationModel.findOne({
-      organizationId: orgCourse.organizationId,
-      apiPlatform: props.apiPlatform,
-    });
-    if (!orgIntegration) {
-      return ERROR_MESSAGES.courseController.orgIntegrationNotFound;
-    }
-
-    const exists = await LMSCourseIntegrationModel.findOne({
-      where: { courseId: courseId, orgIntegration: orgIntegration },
-    });
-    if (!exists) {
-      return ERROR_MESSAGES.courseController.lmsIntegrationNotFound;
-    }
-
-    await LMSCourseIntegrationModel.remove(exists);
-    return `Successfully disconnected LMS integration`;
   }
 }