+++ packages/server/src/course/course.entity.ts
@@ -22,6 +22,7 @@ import { CourseSettingsModel } from './course_settings.entity';
 import { QuestionTypeModel } from '../questionType/question-type.entity';
 import { StudentTaskProgressModel } from '../studentTaskProgress/studentTaskProgress.entity';
 import { LMSCourseIntegrationModel } from '../lmsIntegration/lmsCourseIntegration.entity';
+import { UnreadAsyncQuestionModel } from '../asyncQuestion/unread-async-question.entity';
 @Entity('course_model')
 export class CourseModel extends BaseEntity {
   @PrimaryGeneratedColumn()
@@ -126,6 +127,12 @@ export class CourseModel extends BaseEntity {
     (type) => LMSCourseIntegrationModel,
     (integration) => integration.course,
   )
-  @JoinColumn({ name: 'courseId' })
   lmsIntegration?: LMSCourseIntegrationModel;
+
+  @OneToMany(
+    (type) => UnreadAsyncQuestionModel,
+    (unreadAsyncQuestion) => unreadAsyncQuestion.course,
+  )
+  @Exclude()
+  unreadAsyncQuestions: UnreadAsyncQuestionModel[];
 }