+++ packages/server/test/asyncQuestion.integration.ts
@@ -14,7 +14,7 @@ import { overrideRedisQueue, setupIntegrationTest } from './util/testUtils';
 import { asyncQuestionModule } from 'asyncQuestion/asyncQuestion.module';
 import { AsyncQuestion, asyncQuestionStatus, Role } from '@koh/common';
 import { AsyncQuestionVotesModel } from 'asyncQuestion/asyncQuestionVotes.entity';
-import { UserCourseModel } from 'profile/user-course.entity';
+import { UnreadAsyncQuestionModel } from 'asyncQuestion/unread-async-question.entity';
 
 describe('AsyncQuestion Integration', () => {
   const supertest = setupIntegrationTest(
@@ -72,9 +72,13 @@ describe('AsyncQuestion Integration', () => {
 
   describe('POST asyncQuestions/:cid', () => {
     it('Student can create a question', async () => {
-      const prevCount = await UserCourseModel.findOne({
-        where: { userId: studentUser2.id, courseId: course.id },
-      });
+      const [prevRecords, prevCount] =
+        await UnreadAsyncQuestionModel.findAndCount({
+          where: {
+            userId: studentUser2.id,
+            courseId: course.id,
+          },
+        });
 
       await supertest({ userId: studentUser.id })
         .post(`/asyncQuestions/${course.id}`)
@@ -84,12 +88,14 @@ describe('AsyncQuestion Integration', () => {
         })
         .expect(201)
         .then(async (response) => {
-          const currentCount = await UserCourseModel.findOne({
-            where: { userId: studentUser2.id, courseId: course.id },
-          });
-          expect(currentCount.unreadAsyncQuestions).toBe(
-            prevCount.unreadAsyncQuestions + 1,
-          );
+          const [currentRecords, currentCount] =
+            await UnreadAsyncQuestionModel.findAndCount({
+              where: {
+                userId: studentUser2.id,
+                courseId: course.id,
+              },
+            });
+          expect(currentCount).toBe(prevCount + 1);
 
           expect(response.body).toHaveProperty('status', 'AIAnswered');
           expect(response.body).toHaveProperty('closedAt', null);