+++ packages/frontend/public/changelog.md
@@ -1,3 +1,23 @@
+## Version 1.3.0 - Feb 05, 2025
+
+#### New Features
+- ***General***
+  - Added a new "System" Chatbot toggle, allowing you to ask any questions or feedback regarding the HelpMe system 
+  	- By default, it will still be in "Course" mode, which will query uploaded course documents as before
+  	- HelpMe developers will frequently be looking at these System Chatbot questions to improve the system and Chatbot
+    	- So if you notice that the answer was not accurate, try asking it again in a couple days, it may be better!
+- ***Staff-only***
+  - Fixed and completely overhauled the Learning Management System (Canvas) integration
+  	- Professors can now link their Canvas courses with HelpMe, allowing:
+  		- Professors to synchronize assignments (descriptions) and announcements to the HelpMe course chatbot
+  		- Professors to compare and see which students is in your HelpMe course and not your Canvas course (or vise-versa)
+
+#### Improvements
+- Adjusted and improved tooltips under Chatbot Settings and Export Data (under Course Settings)
+
+
+***
+
 ## Version 1.2.3 - Feb 02, 2025
 
 #### New Features